Shooters, Intake, and Gears-ish
The shooter code is in it's own class, as Cullen suggested. We didn't know where to put the class, however, and it's probably in the wrong spot in terms of organization. Anyway, we tested it and it seemed to be working (Judging from the flashy lights on Talons). We also changed the button for toggling the solenoid we'll use for the gear pistons and added some basic code for activating the intake roller.
- C^2